Added the SparkReplayer monitor.
The spark replayer automatically offloads analysis and decoding to a Spark
cluster. By default, the local pyspark instance is chosen, but this can
be overwritten using the MI_SPARK_URL environment variable to use a
remote cluster. Results from analyzers are collected using collect() on
the analyzers. If pyspark is not available on the local machine, all
spark-related extensions are hidden from import.